Made changes and Fixes to clean the code

+++ src/main/java/com/dining/diningreview/repository/RestaurantRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.dining.diningreview.model.Restaurant;
 
-public interface RestaurantRepository extends CrudRepository<Restaurant, Integer> {
+public interface RestaurantRepository extends CrudRepository<Restaurant, Long> {
     Optional<Restaurant> findRestaurantByNameAndZipcode(String name, String zipcode);
     List<Restaurant> findRestaurantByZipcodeAndPeanutScoreNotNullOrderByPeanutScore(String zipcode);
     List<Restaurant> findRestaurantByZipcodeAndDairyScoreNotNullOrderByDairyScore(String zipcode);
